Balancing Bolts/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
python-multipart==0.0.6
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
pandas==2.1.3
requests==2.31.0
